+++ .github/workflows/pr.yml
@@ -28,6 +28,10 @@ jobs:
         uses: actions/checkout@v3
         with:
           fetch-depth: 1
+      - name: Use Node.js 18.x
+        uses: actions/setup-node@v3
+        with:
+          node-version: 18.x
       - name: Check docs are up to date
         shell: bash
         run: |
