--- docs/code/interfaces/types_debug_utils.ErrnoException.md
@@ -1,105 +1,0 @@
-[@algorandfoundation/algokit-utils](../README.md) / [types/debug-utils](../modules/types_debug_utils.md) / ErrnoException
-
-# Interface: ErrnoException
-
-[types/debug-utils](../modules/types_debug_utils.md).ErrnoException
-
-## Hierarchy
-
-- `Error`
-
-  ↳ **`ErrnoException`**
-
-## Table of contents
-
-### Properties
-
-- [code](types_debug_utils.ErrnoException.md#code)
-- [errno](types_debug_utils.ErrnoException.md#errno)
-- [message](types_debug_utils.ErrnoException.md#message)
-- [name](types_debug_utils.ErrnoException.md#name)
-- [path](types_debug_utils.ErrnoException.md#path)
-- [stack](types_debug_utils.ErrnoException.md#stack)
-- [syscall](types_debug_utils.ErrnoException.md#syscall)
-
-## Properties
-
-### code
-
-• `Optional` **code**: `string`
-
-#### Defined in
-
-[src/types/debug-utils.ts:13](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L13)
-
-___
-
-### errno
-
-• `Optional` **errno**: `number`
-
-#### Defined in
-
-[src/types/debug-utils.ts:12](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L12)
-
-___
-
-### message
-
-• **message**: `string`
-
-#### Inherited from
-
-Error.message
-
-#### Defined in
-
-node_modules/typescript/lib/lib.es5.d.ts:1068
-
-___
-
-### name
-
-• **name**: `string`
-
-#### Inherited from
-
-Error.name
-
-#### Defined in
-
-node_modules/typescript/lib/lib.es5.d.ts:1067
-
-___
-
-### path
-
-• `Optional` **path**: `string`
-
-#### Defined in
-
-[src/types/debug-utils.ts:14](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L14)
-
-___
-
-### stack
-
-• `Optional` **stack**: `string`
-
-#### Inherited from
-
-Error.stack
-
-#### Defined in
-
-node_modules/typescript/lib/lib.es5.d.ts:1069
-
-___
-
-### syscall
-
-• `Optional` **syscall**: `string`
-
-#### Defined in
-
-[src/types/debug-utils.ts:15](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L15)
