+++ docs/code/modules/types_debug_utils.md
@@ -13,86 +13,5 @@
 ### Interfaces
 
 - [AVMDebuggerSourceMapDict](../interfaces/types_debug_utils.AVMDebuggerSourceMapDict.md)
-- [ErrnoException](../interfaces/types_debug_utils.ErrnoException.md)
 - [PersistSourceMapsParams](../interfaces/types_debug_utils.PersistSourceMapsParams.md)
 - [SimulateAndPersistResponseParams](../interfaces/types_debug_utils.SimulateAndPersistResponseParams.md)
-
-### Variables
-
-- [ALGOKIT\_DIR](types_debug_utils.md#algokit_dir)
-- [DEBUG\_TRACES\_DIR](types_debug_utils.md#debug_traces_dir)
-- [SOURCES\_DIR](types_debug_utils.md#sources_dir)
-- [SOURCES\_FILE](types_debug_utils.md#sources_file)
-- [TEAL\_FILE\_EXT](types_debug_utils.md#teal_file_ext)
-- [TEAL\_SOURCEMAP\_EXT](types_debug_utils.md#teal_sourcemap_ext)
-- [TRACES\_FILE\_EXT](types_debug_utils.md#traces_file_ext)
-
-## Variables
-
-### ALGOKIT\_DIR
-
-• `Const` **ALGOKIT\_DIR**: ``".algokit"``
-
-#### Defined in
-
-[src/types/debug-utils.ts:3](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L3)
-
-___
-
-### DEBUG\_TRACES\_DIR
-
-• `Const` **DEBUG\_TRACES\_DIR**: ``"debug_traces"``
-
-#### Defined in
-
-[src/types/debug-utils.ts:7](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L7)
-
-___
-
-### SOURCES\_DIR
-
-• `Const` **SOURCES\_DIR**: ``"sources"``
-
-#### Defined in
-
-[src/types/debug-utils.ts:4](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L4)
-
-___
-
-### SOURCES\_FILE
-
-• `Const` **SOURCES\_FILE**: ``"sources.avm.json"``
-
-#### Defined in
-
-[src/types/debug-utils.ts:5](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L5)
-
-___
-
-### TEAL\_FILE\_EXT
-
-• `Const` **TEAL\_FILE\_EXT**: ``".teal"``
-
-#### Defined in
-
-[src/types/debug-utils.ts:8](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L8)
-
-___
-
-### TEAL\_SOURCEMAP\_EXT
-
-• `Const` **TEAL\_SOURCEMAP\_EXT**: ``".teal.tok.map"``
-
-#### Defined in
-
-[src/types/debug-utils.ts:9](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L9)
-
-___
-
-### TRACES\_FILE\_EXT
-
-• `Const` **TRACES\_FILE\_EXT**: ``".trace.avm.json"``
-
-#### Defined in
-
-[src/types/debug-utils.ts:6](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L6)
